+++ Package.swift
@@ -14,7 +14,8 @@ let package = Package(
 	targets: [
 		.binaryTarget(
 			name: "CContextDB",
-			path: "dist/ContextDB.xcframework"
+			url: "https://github.com/charliewilco/ContextDB/releases/download/v0.1.0/ContextDB.xcframework.zip",
+			checksum: "2ae40df138e61416207382df471a2084b534e38a71cd4f9cef7728f9930553a8"
 		),
 		.target(
 			name: "ContextDB",